--- a/Tests/JsonReader_Read_Tests.cs
+++ b/Tests/JsonReader_Read_Tests.cs
@@ -216,9 +216,38 @@ namespace Tests
 
             // asert Current
             Assert.False(jsonReader.ReadAsBoolean());
-
-
         }
 
+        [Fact]
+        public void ReadAsEscapedString_ShouldReturnEscapedString()
+        {
+            const string jsonEscapedString = "\"Hello\\nWorld\"";
+
+            using var stream = new MemoryStream(Encoding.UTF8.GetBytes(jsonEscapedString));
+            var jsonReader = new JsonReader(stream, bufferSize: 10);
+
+            jsonReader.Read();
+            Assert.Equal(JsonTokenType.String, jsonReader.TokenType);
+            var result = jsonReader.GetEscapedString();
+
+            Assert.Equal("Hello\\nWorld", result);
+        }
+
+        [Fact]
+        public void ReadAsString_ShouldReturnUnescapedString()
+        {
+            const string jsonEscapedString = "\"Hello\\n\\u003EWorld\"";
+
+            using var stream = new MemoryStream(Encoding.UTF8.GetBytes(jsonEscapedString));
+            var jsonReader = new JsonReader(stream, bufferSize: 10);
+
+            jsonReader.Read();
+            Assert.Equal(JsonTokenType.String, jsonReader.TokenType);
+            var result = jsonReader.GetString();
+
+            const string jsonUnescapedString = "Hello\n>World";
+
+            Assert.Equal(jsonUnescapedString, result);
+        }
     }
 }