Reduces allocations. Added some tests.

Before:

| Method | Mean     | Error    | StdDev   | Gen0       | Allocated |
|------- |---------:|---------:|---------:|-----------:|----------:|
| Read   | 20.49 ms | 0.313 ms | 1.276 ms |  3000.0000 |  27.49 MB |
| Values | 54.52 ms | 0.496 ms | 2.079 ms | 15000.0000 | 140.59 MB |

After:

| Method | Mean     | Error    | StdDev   | Gen0       | Allocated |
|------- |---------:|---------:|---------:|-----------:|----------:|
| Read   | 19.04 ms | 0.261 ms | 1.081 ms |  2000.0000 |  20.01 MB |
| Values | 54.25 ms | 0.666 ms | 2.805 ms | 14000.0000 | 133.11 MB |



-----

If I change the Read benchmark to use `GetEscapedString()` instead of `GetString()` we see the following results




Before:

| Method | Mean     | Error    | StdDev   | Gen0       | Allocated |
|------- |---------:|---------:|---------:|-----------:|----------:|
| Read   | 20.21 ms | 0.257 ms | 1.047 ms |  3000.0000 |  27.49 MB |

After:

| Method | Mean     | Error    | StdDev   | Gen0       | Allocated |
|------- |---------:|---------:|---------:|-----------:|----------:|
| Read   | 18.87 ms | 0.269 ms | 1.108 ms |  2000.0000 |  20.01 MB |