@Mimetis 2024-05-08
good catch using `.Span` !